Split prior weights modules and unify aggregate spec model.
CountsAggregate uses pseudo_count; registry lookup is cached; fixed and template
entries share one AggregateActionSpec and a single catalog slot append path.
